# Conflicts:
#	.flutter-plugins-dependencies
#	build_variant.yaml
#	lib/background/background_sync.dart
#	lib/logger/logger.dart
#	lib/modules/auth/blocs/auth_bloc/auth_methods.dart
#	lib/modules/mail/blocs/mail_bloc/mail_methods.dart
#	pubspec.yaml

+++ lib/modules/auth/blocs/auth_bloc/auth_methods.dart
@@ -123,8 +123,8 @@ class AuthMethods {
 
     final futures = [
       deleteUserRelatedData(user),
-      if(user.localId==currentUserId)
-      _cryptoStorage.deleteAll(),
+      if (_cryptoStorage != null && user.localId==currentUserId)
+        _cryptoStorage.deleteAll(),
       _authLocal.deleteSelectedUserLocalId(),
       if(user.localId==currentUserId)
       _authLocal.deleteSelectedAccountId(),
